Loadable sections
So far, there was a list of types of sections, that were supposed
to be loaded into memory. This method is now replaced with a section
flag "loadable" - if this flag is set, section is loaded, no questions
asked. By default, .text, .data, .bss sections are loadable, others
are not. User-defined sections can use it, too, by setting 'l' flag
in .section directive.