Pylint changes
Pylint got updated a lot since I used it for the last time,
so I updated its rcfile, and added make target for it. My code
needs a bit of polishing to get rid of so far huge pile of
warnings, and later clean sheet will be a prerequisite to
a commit, like flake is already.