Going 32-bits! And FPGA! \o/
For some reason, I got an idea - why not to take a FPGA, and create a real
Ducky CPU? Unfortunatelly, the current state of Ducky wasn't very hardware
friendly, and I was getting quite upset about the segmentation itself, so I
decided to make few tiny, tiny changes...

This huge (yes, it has more lines than the code itself) patch bring a lot of
fundamental changes, as well as a one-liner fixes of issues I discovered
during the coding. There are still few issues to be solved in the future,
e.g. the `ANS` test suite does not pass yet, not even to the extend before
this patch, and compiler unit tests are not fixed to match new compiler
output, but all in all, it is possible to run programs as complex as `FORTH`
kernel. The VHDL experiments will follow soon, I already got my very first
FPGA \o/

  - a CPU core is now fully 32-bit - registers, address bus, operands...
  - segment registers are gone, every single byte of memory is addresable
  - directly
  - reconstructed instruction encoding
    - reduced number of opcodes, by merging instructions and differentiating
    - by flags
    - reduced number of different encodings, to simplify decoder hardware
    - design
    - jump destinations (`j`, `b*` and `call` instructions) must be always
    - aligned to 4 byte boundary, therefore the values encoded in
    - instructions are shifted by 2 to the right, and shifted again - to the
    - left - before using.

Reimplemented to provide more hardware-like feel - this way it will be
easier to keep in sync with any hardware implementation of Ducky.

When powered on, CPU core sets its `IP` to a hard-coded address
(`0x00020000` by default). It expects there's a stream of valid instruction
starting at this address. I suppose that, in case of FPGA, I can hardcore
bootloader (right now cca 688 bytes) directly into the core implementation.
Therefore I need a image of bootloader's binary file that I could encode
into `VHDL`, and I need a way to load such image - or binary - when I use
VM. VM's `ROMLoader` module can do a bit more but in general it provides the
requested functionality.

Bootloader then sets minimal environment - esecially with respect to SMP -
and then, using one of available methods, loads the next stage of booting
process, the binary. The idea is to have a single source for a bootloader,
with selectable methods of loading the next stage, compiled in on demand.
The resulting bootloader binary would then be either translated to an image,
ready for encoding into FPGA implementation, or loaded by `ROMLoader` of VM.
Currently, only one method is available (loading from VM's storage device),
but others - like FLASH memories or some forms of `SPI` buses.

It works well, as can be seen in `smp` example.

  - block IO subsystem reimplemented to resemble interfaces of the real
  - block devices, similar to `ATAPI`. So far `DMA` mode is supported only,
  - I'd like to add `PIO` mode later.
  - terminals support echo, but it's only possible to change it in
  - configuration file, not during runtime
  - math coprocessor operates on 64-bit integers
  - math coprocessor does not provide virtual interrupt anymore
  - control coprocessor does not have its own instructions, I inserted the
  - basic instructions, `CTR` and `CTW`, into the main instruction set

  - internal event bus - used so far just for signals from CPU cores to VM
  - when changing their runnable state
  - shorter names for `ctypes` types, e.g. `UInt32` is now `u32_t`. It also
  - resembles glibc's data types, reminding the reader that this variable is
  - sensitive
  - IO streams can provide `flush()` method
  - VM does no symbol resolving anymore
  - `interrupts` file is no longer necessary
  - commonly used `Flags` class no longer depends on `ctypes` - it was not
  - necessary for its functionality, by adding conversion methods it is
  - possible to save and restore its state in binary files or in memory
  - there is no default stack or `PT` at the beginning of execution - if
  - code wants to use stack, it must prepare its own
  - debugging points are now independed on the following action - it is
  - possible to bind different points to the same actions, or multiple
  - actions to a single point

  - `.align` directive - this one I was missing a lot, now it's easy to
  - align symbols. It proved to be extremely useful in porting `FORTH`
  - kernel
  - `.global` directive - symbols marked as *global* can be referenced from
  - other other object files when linking
  - `.short` data type - since this patch, the integer (derived from `word`)
  - is now 4 bytes long. The previous `int`, 2 bytes long, becomes `short`
  - I've been using `ctypes`-based variables in assembler, for several
  - purposes, when it was not necessary - standard Python integer is good
  - enough, so I removed `ctypes` from the code flow where possible. Values
  - are then converted correctly when object file is being written.
  - internal handling of references to symbols was been very scattered and
  - sometimes burried deep in code (remnants of several changes of my
  - mind...). This area is now much cleaner, using objects to store
  - necessary information.
  - during the encoding of instructions, overflow of immediate values can be
  - detected, and is reported
  - `--verify-disassemble` option - when set, it disassembles currently
  - assembled instruction, and compares it to its model in the source file.
  - It has some flaws but for sanity verification - especially after some
  - changes to instruction encoding - it has proven to be very useful.
  - assembly defines
    - moved to a separate directory
    - generated from templates, using `Mako` template engine, to keep their
    - content in sync with sources

  - like assembler, linker also checks for situations where immediate
  - overflows space it is supposed to patch during relocation process

  - `ducky-image` is a new tool that can create binary images, later used by
  - Ducky VM or in its hardware implementation
  - it is also possible to create images of `HDT`, for later encoding into
  - hardware implementation of CPU

  - we have 32-bit CPU, we must have a 32-bit `FORTH` kernel then...
  - kernel is loaded either by bootloader, or by VM's ROM loader - in both
  - scenarios, it is located at the *wrong* address. Because of the way how
  - dictionary of words is linked together, by absolute addresses stored in
  - `.` and `link` assembler variables, it is not possible to run kernel at
  - this location, therefore the kernel moves itself to the beggining of the
  - address space.
  - `FORTH_TEXT_WRITABLE` option deprecated and removed since it makes not
  - much sense after this patch
  - double-cell integer words moved into separate file - I intend to split
  - kernel source to more files, I don't like huge piles of characters...
  - ANS test suite is now bundled as a git submodule - it is now on Github,
  - this way it will be always actual
  - ANS test suite already revealed few issues with core words, e.g. `WORD`
  - according to a standard expectes delimiter character on the stack - my
  - implementation was not accurate

And definitely some others...