[Linker] Duplicity check dropped for a better algorithm
Duplicity check kept track of symbols with the same name, defined in
multiple files, which could confuse linker when it's supposed to find
the correct symbol referenced by relocation entry. Now it gives priority
to those symbols that origin in the same object file as a place that
references them.